+++ spreadsheet-api/src/test/java/ec/util/spreadsheet/BookTest.java
@@ -69,7 +69,7 @@ public class BookTest {
                             }
 
                             @Override
-                            public String getName() {
+                            public @org.jspecify.annotations.NonNull String getName() {
                                 return "hello";
                             }
                         };