+++ spreadsheet-markdown/src/main/java/module-info.java
@@ -1,0 +1,12 @@
+module nbbrd.spreadsheet.markdown {
+
+    requires static org.jspecify;
+    requires static nbbrd.service;
+    requires static lombok;
+
+    requires nbbrd.spreadsheet.api;
+
+    provides ec.util.spreadsheet.Book.Factory with
+            ec.util.spreadsheet.markdown.MarkdownBookFactory;
+}
+