+++ spreadsheet-poi/src/test/java/ec/util/spreadsheet/poi/ExcelBookFactoryTest.java
@@ -128,10 +128,10 @@ public class ExcelBookFactoryTest {
 
     @Test
     public void testRank() {
-        assertThat(BookFactoryLoader.get())
-                .hasSize(2)
+        assertThat(BookFactoryLoader.load())
+                .hasSize(3)
                 .isSortedAccordingTo(Comparator.comparingInt(Book.Factory::getRank).reversed())
                 .extracting(factory -> factory.getClass().getSimpleName())
-                .containsExactly("ExcelBookFactory", "ExcelClassicBookFactory");
+                .containsExactly("ExcelBinaryBookFactory", "ExcelBookFactory", "ExcelClassicBookFactory");
     }
 }