+++ spreadsheet-xl/src/test/java/spreadsheet/xlsx/XlsxBookFactoryTest.java
@@ -121,10 +121,10 @@ public class XlsxBookFactoryTest {
 
     @Test
     public void testRank() {
-        assertThat(BookFactoryLoader.get())
-                .hasSize(3)
+        assertThat(BookFactoryLoader.load())
+                .hasSize(4)
                 .isSortedAccordingTo(Comparator.comparingInt(Book.Factory::getRank).reversed())
                 .extracting(factory -> factory.getClass().getSimpleName())
-                .containsExactly("XlsxBookFactory", "ExcelBookFactory", "ExcelClassicBookFactory");
+                .containsExactly("XlsxBookFactory", "ExcelBinaryBookFactory", "ExcelBookFactory", "ExcelClassicBookFactory");
     }
 }